renamed files and fixed paths
