+++ .gitignore
@@ -19,10 +19,9 @@ Thumbs.db
 *.sw?
 *.vi
 
-/node_modules/
+node_modules/
 npm-debug.log
+lib/
 demo/bundle.js
-/lib
-/native.js
 coverage
 addons/
