+++ demo/app.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import look, { StyleSheet, Presets } from '../modules/look'
+import look, { StyleSheet, Presets } from '../modules/dom'
 import Container from './components/layout/Container.jsx'
 import UserAction from './components/examples/UserAction.jsx'
 import ChildIndex from './components/examples/ChildIndex.jsx'
@@ -12,46 +12,42 @@ import Pseudo from './components/examples/Pseudo.jsx'
 
 const App = () => (
 <div>
-		<h1 className={styles}>Look Examples</h1>
-		<Container group title="1. Pseudo classes">
-			<Container title="1.1. user action">
-				<UserAction />
-			</Container>
-			<Container noPadding title="1.2. child index & child type">
-				<ChildIndex items={14} />
-			</Container>
-			<Container title="1.3. input">
-				<Input />
-			</Container>
-      <Container title="1.4. other">
-        <Pseudo />
-      </Container>
-		</Container>
-
-		<Container group title="2. Stateful conditions">
-			<Container description="Styles depend on props, state & context">
-				<Condition />
-			</Container>
-		</Container>
-
-		<Container group title="3. Media queries">
-			<Container noPadding>
-				<Media />
-			</Container>
-		</Container>
-
-    <Container group title="4. Platform queries">
-      <Container noPadding>
-        <Platform />
-      </Container>
+  <h1 className={styles}>Look Examples</h1>
+  <Container group title="1. Pseudo classes">
+    <Container title="1.1. user action">
+      <UserAction />
     </Container>
-
-    <Container group title="5. Keyframe animations">
-      <Container noPadding description="Using the StyleSheet.keyframes helper">
-        <Keyframes />
-      </Container>
+    <Container noPadding title="1.2. child index & child type">
+      <ChildIndex items={14} />
     </Container>
-	</div>
+    <Container title="1.3. input">
+      <Input />
+    </Container>
+    <Container title="1.4. other">
+      <Pseudo />
+    </Container>
+  </Container>
+  <Container group title="2. Stateful conditions">
+    <Container description="Styles depend on props, state & context">
+      <Condition />
+    </Container>
+  </Container>
+  <Container group title="3. Media queries">
+    <Container noPadding>
+      <Media />
+    </Container>
+  </Container>
+  <Container group title="4. Platform queries">
+    <Container noPadding>
+      <Platform />
+    </Container>
+  </Container>
+  <Container group title="5. Keyframe animations">
+    <Container noPadding description="Using the StyleSheet.keyframes helper">
+      <Keyframes />
+    </Container>
+  </Container>
+</div>
 )
 
 const styles = StyleSheet.create({
