+++ demo/components/examples/ChildIndex.jsx
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from 'react'
-import look, { StyleSheet } from '../../../modules/look'
+import look, { StyleSheet } from '../../../modules/dom'
 
 class ChildIndex extends Component {
   static defaultProps = {
@@ -9,10 +9,13 @@ class ChildIndex extends Component {
     items: PropTypes.number.isRequired
   };
   render() {
-    //create a small list of items to demonstrate nth-child, et cetera
-    const arr = new Array(this.props.items+1).join(',').split('').map((item, index) => index + 1)
+    // create a small list of items to demonstrate nth-child, et cetera
+    const arr = new Array(this.props.items + 1).join(',').split('').map((item, index) => index + 1)
     let list = arr.map((text, index) => {
-      return <li className={styles} key={'.' + index}>ListItem {text}</li>
+      return <li className={styles} key={'.' + index}>
+               ListItem&nbsp;
+               {text}
+             </li>
     })
 
     return (
@@ -26,24 +29,24 @@ class ChildIndex extends Component {
 }
 
 const styles = StyleSheet.create({
-    color: 'darkgray',
-    listStyleType: 'none',
-    ':nth-child(odd)': {
-      backgroundColor: 'lightGray',
-      color: 'blue'
-    },
-    ':last-child': {
-      color: 'red'
-    },
-    ':nth-of-type(3)': {
-      backgroundColor: 'rgba(30, 50, 60, 0.6)'
-    },
-    ':nth-of-type(3n+5)'  :{
-      backgroundColor:'rgba(141, 195, 131, 0.4)'
-    },
-    ':first-child': {
-      color: 'yellow'
-    }
+  color: 'darkgray',
+  listStyleType: 'none',
+  ':nth-child(odd)': {
+    backgroundColor: 'lightGray',
+    color: 'blue'
+  },
+  ':last-child': {
+    color: 'red'
+  },
+  ':nth-of-type(3)': {
+    backgroundColor: 'rgba(30, 50, 60, 0.6)'
+  },
+  ':nth-of-type(3n+5)': {
+    backgroundColor: 'rgba(141, 195, 131, 0.4)'
+  },
+  ':first-child': {
+    color: 'yellow'
+  }
 })
 
 export default look(ChildIndex)
