+++ demo/components/examples/Condition.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import look, { StyleSheet } from '../../../modules/look'
+import look, { StyleSheet } from '../../../modules/dom'
 const c = StyleSheet.combineStyles
 
 import Color from 'color'
@@ -41,10 +41,16 @@ class Condition extends Component {
 
     return (
       <div>
-        <div className={c(styles.button, styles.states)} onClick={this.onClick.bind(this)}>Click me<br /> Active state: {this.state.mode}</div>
-      <div className={c(styles.button, styles.clicks)} onClick={this.onClicksCount.bind(this)}>{getText(this.state.clicks)}</div>
+        <div className={c(styles.button, styles.states)} onClick={this.onClick.bind(this)}>
+          Click me
+          <br /> Active state:
+          {this.state.mode}
+        </div>
+        <div className={c(styles.button, styles.clicks)} onClick={this.onClicksCount.bind(this)}>
+          {getText(this.state.clicks)}
+        </div>
       </div>
-      )
+    )
   }
 }
