+++ demo/components/examples/Keyframes.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import look, { StyleSheet } from '../../../modules/look'
+import look, { StyleSheet } from '../../../modules/dom'
 
 const Keyframes = () => (
 <div className={styles.container}>
