+++ demo/components/examples/Platform.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import look, { StyleSheet } from '../../../modules/look'
+import look, { StyleSheet } from '../../../modules/dom'
 const c = StyleSheet.combineStyles
 
 const iOS = 'Mozilla/5.0 (iPhone; CPU iPhone OS 7_0 like Mac OS X) AppleWebKit/538.34.9 (KHTML, like Gecko) Mobile/12A4265u'
