+++ demo/components/layout/Container.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import look, { StyleSheet } from '../../../modules/look'
+import look, { StyleSheet } from '../../../modules/dom'
 
 class Container extends Component {
   render() {
