+++ demo/server.js
@@ -4,7 +4,7 @@ import React from 'react'
 import { renderToString } from 'react-dom/server'
 import App from './app.jsx'
 import fs from 'fs'
-import { Presets, LookRoot } from '../modules/look'
+import { Presets, LookRoot } from '../modules/dom'
 
 const indexHTML = fs.readFileSync(__dirname + '/index.html').toString()
 const app = express()
