+++ modules/dom.js
@@ -1,9 +1,13 @@
 // Plugins
+import fallbackValue from './plugins/fallbackValue'
 import statefulValue from './plugins/statefulValue'
 import mixin from './plugins/mixin'
+import prefixer from './plugins/prefixer'
 
 // Dev tools
+import linter from './plugins/linter'
 import styleLogger from './plugins/styleLogger'
+import friendlyClassName from './plugins/friendlyClassName'
 
 // Conditions
 import { equal, unEqual, greater, less, greaterThan, lessThan } from './mixins/condition'
@@ -11,14 +15,17 @@ import { equal, unEqual, greater, less, greaterThan, lessThan } from './mixins/c
 // Mixins
 import contains from './mixins/contains'
 import substr from './mixins/substr'
+import extractCSS from './mixins/extractCSS'
 import extend from './mixins/extend'
+import platformQuery from './mixins/platformQuery'
 
 // Presets
 import nativePreset from './presets/react-native'
+import domPreset from './presets/react-dom'
 
 import Enhancer from './core/enhancer'
-import StyleSheet from './native/api/StyleSheet'
-import LookRoot from './native/api/LookRoot'
+import StyleSheet from './dom/api/StyleSheet'
+import LookRoot from './dom/api/LookRoot'
 
 // Resolving annotations
 // If not passing arguments it just wraps the Component
@@ -36,8 +43,12 @@ export default (...args) => {
 
 const Plugins = {
   mixin,
+  fallbackValue,
   statefulValue,
-  styleLogger
+  prefixer,
+  styleLogger,
+  linter,
+  friendlyClassName
 }
 
 const Mixins = {
@@ -51,12 +62,14 @@ const Mixins = {
 
   // Other
   extend,
+  extractCSS,
+  platformQuery,
   contains,
   substr
 }
 
 const Presets = {
-  'react-native': nativePreset
+  'react-dom': domPreset
 }
 
 export {
