--- modules/look.js
@@ -1,29 +1,0 @@
-import Enhancer from './core/enhancer'
-import StyleSheet from './dom/api/StyleSheet'
-import LookRoot from './dom/api/LookRoot'
-
-import NativeStyleSheet from './native/api/StyleSheet'
-
-import { Plugins, Mixins, Presets } from './addons'
-
-// Resolving annotations
-// If not passing arguments it just wraps the Component
-// Otherwise it returns a decorator
-export default (...args) => {
-  if (args[0] instanceof Function) {
-    return Enhancer(...args) // eslint-disable-line
-  }
-
-  return function decorator(target) {
-    return Enhancer(target, ...args) // eslint-disable-line
-  }
-}
-
-export {
-  StyleSheet,
-  LookRoot,
-
-  Plugins,
-  Mixins,
-  Presets
-}
